// src/features/draft/commands/types/api-remove-rows.handler.types.ts
import { GetBranchByIdReturnType } from 'src/features/branch/quieries/types/get-branch-by-id.types';
import { GetTableByIdReturnType } from 'src/features/table/queries/types';

export type ApiRemoveRowsHandlerReturnType = {
  branch: GetBranchByIdReturnType;
  table?: GetTableByIdReturnType;
  previousVersionTableId?: string;
};